+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet.
+
+## [v2.0.1] - 2022-12-07
+
 ### Added
 
 - `float` type in `put` command.
@@ -515,7 +519,8 @@ See [documentation](https://daseldocs.tomwright.me) for all changes.
 
 - Everything!
 
-[unreleased]: https://github.com/TomWright/dasel/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/TomWright/dasel/compare/v2.0.1...HEAD
+[v2.0.1]: https://github.com/TomWright/dasel/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/TomWright/dasel/compare/v1.27.3...v2.0.0
 [v1.27.3]: https://github.com/TomWright/dasel/compare/v1.27.2...v1.27.3
 [v1.27.2]: https://github.com/TomWright/dasel/compare/v1.27.1...v1.27.2